requirements.txt
Flask
Flask-cors

requests
pyotp
numpy
pytz
robin_stocks
telepot
python-dotenv
json